requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
transformers
sentence-transformers
pymongo
llama-cpp-python